refactor: rebase the project
We project to add more applications in the future.
For it, organized the project in packages

Based on the demand to have the same afrim web bahavior, we separed the
afrim website and afrim input in writing a library that anyone can use.

As additional change, we updated some dependencies and added
minor updates.